♻️ refac: app.js와 page 분리
