+++ src/scss/MoveBtn.scss
@@ -1,4 +1,4 @@
-@import "App.scss";
+@import "Main.scss";
 
 .MoveBtn-flex {
   display: flex;
