+++ src/scss/Settings.scss
@@ -1,4 +1,4 @@
-@import "App.scss";
+@import "Main.scss";
 
 @mixin setting-box {
   border-bottom: 1px solid $light-grey;
