+++ src/scss/dualSelector.scss
@@ -1,4 +1,4 @@
-@import "App.scss";
+@import "Main.scss";
 
 .dual-selector-wrap {
   background-color: #fff;
